move json write funciton to analysis as it is a specific function
